refactor: break up 3 remaining large view files into focused sub-views (#603)
Completes the work started in PR #608. Breaks up the 3 files that were
still over 800 LOC into focused extensions:

- RulesSummaryView.swift: 1,102 → 663 (3 new extension files)
- BaseKeycap.swift: 850 → 70 (3 new extension files)
- ContextHUDController.swift: 863 → 332 (2 new extension files)
